[REFACTOR] Service create command (#234)
* [REFACTOR] Service create command

* remove failing test

* remove unwanted files

* incorporated comments from review

* update code based on review feedback

* remove try catch block around Bedrock function call

+++ src/lib/ioUtil.ts
@@ -24,15 +24,17 @@ export const createTempDir = (parent?: string): string => {
  */
 export const removeDir = (dir: string) => {
   const folder = path.resolve(dir);
-  fs.readdirSync(folder).forEach(item => {
-    const curPath = path.join(folder, item);
-    if (fs.statSync(curPath).isDirectory()) {
-      removeDir(curPath);
-    } else {
-      fs.unlinkSync(curPath);
-    }
-  });
-  fs.rmdirSync(folder);
+  if (fs.existsSync(folder)) {
+    fs.readdirSync(folder).forEach(item => {
+      const curPath = path.join(folder, item);
+      if (fs.statSync(curPath).isDirectory()) {
+        removeDir(curPath);
+      } else {
+        fs.unlinkSync(curPath);
+      }
+    });
+    fs.rmdirSync(folder);
+  }
 };
 
 /**
